Explain what type errors have to do with purity
To someone who isn't yet familiar with Haskell's syntax, the string
concatenation example looks like the type error is that a string is
being concatenated with a function. Here is an example which better
explains how using an IO computation in a pure context can lead to a
type error.

As a bonus, the new text no longer claims that "you cannot mix and match
purity with impurity". It's not true: we can mix them, but the result
must be marked as impure.